requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9
transformers>=4.30
pandas
Pillow
opencv-python
tqdm
matplotlib
numpy
